// functions/src/types/HospitalServiceType.ts
enum HospitalServiceType {
  LEVEL_1 = 'Level 1',
  LEVEL_2 = 'Level 2',
  LEVEL_3 = 'Level 3',
  LEVEL_1_SPECIALTY = 'Level 1-Specialty',
  LEVEL_2_SPECIALTY = 'Level 2-Specialty',
  LEVEL_3_SPECIALTY = 'Level 3-Specialty',
  SPECIALTY_MOTHER_CHILD = 'Specialty-Mother & Child',
  SPECIALTY_OB_GYNECOLOGY = 'Specialty-OB & Gynecology'
}

export default HospitalServiceType;